+++ NEWS.md
@@ -4,6 +4,7 @@
 *   Fixed a bug with determining the full name of the user running the report, and added the username to the report as well.
 *   Minor tweaks related to the label attributes. Fixes #21 - thanks to jomuller.
 *   Added data.frame label to the presentation under the summary table. Fixes #23 - thanks to carlbfrederick
+*   Fixec a bug with the arguments to makeCodebook - thanks to Jaeoc (#24)
 
 # dataMaid 1.1.0